I should commit more often!
Using argparse, making it way more functional. That was easy.

Then I started learning a lot about how creation time works: it doesn't exist on most UNIX systems, but does on Free BSD and macOS. Because it doesn't exist on most UNIX systems, Python's os library can't handle that function. Come to learn, my old system (bash, touch, and lots of Excel) was only incidentally handling this issue because when touching a file, if the modified date is older than the creation date macOS automatically sets the creation date to the modified date. Some errors earlier meant the creation date was incorrect, but could only be pushed further into the past. So I found macOS's SetFile command, which exists with the installation of developer tools, and added a flag to use that. It's way, way slower, but it works as advertised.